Break out breakdown metrics into its own composable
